(split)Fixed bug in autoclimate based on laston. See below.
There was a logic bug in autoclimate. If a climate was on, then
went offline, then the "turn off" timer signaled, and THEN it went
online again, it would think that laston > turnoff time, and it would
not turn on. (It thinks you manually turned it on.)

I changed "laston" sensors to be the last time it was TURNED on
(ie: off to on).